+++ src/pybricks/robotics.py
@@ -26,12 +26,19 @@ class DriveBase:
     distance_control = _Control()
     """The traveled distance and drive speed are controlled by a PID
     controller. You can use this attribute to change its settings.
-    See :ref:`control` for an overview of available methods."""
+    See the :ref:`motor control <settings>` attribute for an overview of
+    available methods. The ``distance_control`` attribute has the same
+    functionality, but the settings apply to every millimeter driven by the
+    drive base, instead of degrees turned by one motor."""
 
     heading_control = _Control()
     """The robot turn angle and turn rate are controlled by a PID
     controller. You can use this attribute to change its settings.
-    See :ref:`control` for an overview of available methods."""
+    See the :ref:`motor control <settings>` attribute for an overview of
+    available methods. The ``heading_control`` attribute has the same
+    functionality, but the settings apply to every degree of rotation of the
+    whole drive base (viewed from the top) instead of degrees turned by one
+    motor."""
 
     def __init__(self, left_motor, right_motor, wheel_diameter, axle_track):
         """DriveBase(left_motor, right_motor, wheel_diameter, axle_track)