+++ doc/main/CHANGELOG.md
@@ -3,8 +3,11 @@
 ## Added:
 - MicroPython module documentation.
 - Examples for hub system functions including stop button and shutdown.
+
+## Changed:
 - Build IDE docs as main docs with minor changes, instead of a completely
   separate build.
+- Moved motor control documentation to the motor page.
 
 # 1.5.0 - 2021-07-01